// game-app/src/app/shared/model/power.model.ts
export class Power{
    id: number = 0;
    name: string = "";
    description: string = "";
    icon_name: string = "";
    physical_attack: number = 0;
    magic_attack: number = 0;
    physical_defense: number = 0;
    magic_defense: number = 0;
    stamina_consumption: number = 0;
    magic_consumption: number = 0;
    duration_shifts: number = 0;
    shifts_off: number = 0;
}